+++ packages/skin-anocca/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @rttui/skin-anocca
 
+## 1.0.26
+
+### Patch Changes
+
+- Updated dependencies
+  - @rttui/core@1.0.25
+
 ## 1.0.25
 
 ### Patch Changes